Backend/requirements.txt
gitpython
radon
pylint
ast
python-dotenv
lcom
pyyaml
